requirements.txt
yt_dlp==2022.9.1
spotipy==2.16.1
PyInquirer==1.0.3
pyfiglet==0.7
eyed3==0.9.6
python-dotenv==0.21.0
